requirements.txt
pymysql
flask
django-environ
